m12-cognito-userpool-auth/api/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=42.0.0
